% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arReceivable.R
\name{dms_ar_receivable_input_delete}
\alias{dms_ar_receivable_input_delete}
\title{清空rds_dms_ods_t_ar_receivable_input}
\usage{
dms_ar_receivable_input_delete(dms_token)
}
\arguments{
\item{dms_token}{}
}
\value{
无返回值
}
\description{
清空rds_dms_ods_t_ar_receivable_input
}
\examples{
dms_ar_receivable_input_delete()
}
