% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arReceivable.R
\name{dms_ar_receivable_upload}
\alias{dms_ar_receivable_upload}
\title{上传DMS表头表数据}
\usage{
dms_ar_receivable_upload(dms_token)
}
\arguments{
\item{dms_token}{}
}
\value{
无返回值
}
\description{
上传DMS表头表数据
}
\examples{
dms_ar_receivable_upload()
}
