% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arReceivable.R
\name{dms_ar_receivable_view}
\alias{dms_ar_receivable_view}
\title{查询dms 表头更新状态}
\usage{
dms_ar_receivable_view(dms_token, FBillNo, FStartDate, FEndDate)
}
\arguments{
\item{dms_token}{}

\item{FEndDate}{}
}
\value{
无返回值
}
\description{
查询dms 表头更新状态
}
\examples{
dms_ar_receivable_view()
}
