% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arReceivable.R
\name{dms_ar_receivableEntry_input_delete}
\alias{dms_ar_receivableEntry_input_delete}
\title{rds_dms_ods_t_ar_receivableEntry_input清空dms临时表}
\usage{
dms_ar_receivableEntry_input_delete(dms_token)
}
\arguments{
\item{dms_token}{}
}
\value{
无返回值
}
\description{
rds_dms_ods_t_ar_receivableEntry_input清空dms临时表
}
\examples{
dms_ar_receivableEntry_input_delete()
}
