% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outStockMerge.R
\name{dms_sal_outStock_input_delete}
\alias{dms_sal_outStock_input_delete}
\title{清空rds_dms_ods_t_sal_outStock_input}
\usage{
dms_sal_outStock_input_delete(dms_token)
}
\arguments{
\item{dms_token}{}
}
\value{
无返回值
}
\description{
清空rds_dms_ods_t_sal_outStock_input
}
\examples{
dms_sal_outStock_input_delete()
}
