% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salOutStock.R
\name{dms_salOutStockEntry_select}
\alias{dms_salOutStockEntry_select}
\title{查询DMS,salOutStockEntry}
\usage{
dms_salOutStockEntry_select(dms_token, FBillNo)
}
\arguments{
\item{dms_token}{}

\item{FBillNo}{}
}
\value{
无返回值
}
\description{
查询DMS,salOutStockEntry
}
\examples{
dms_salOutStockEntry_select()
}
