% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arReceivable.R
\name{erp_arReceivable_select}
\alias{erp_arReceivable_select}
\title{查询ERP表头input数据}
\usage{
erp_arReceivable_select(erp_token)
}
\arguments{
\item{erp_token}{}
}
\value{
无返回值
}
\description{
查询ERP表头input数据
}
\examples{
erp_arReceivable_select()
}
