% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arReceivable.R
\name{erp_arReceivableEntry_select}
\alias{erp_arReceivableEntry_select}
\title{查询ERP表体input}
\usage{
erp_arReceivableEntry_select(erp_token)
}
\arguments{
\item{erp_token}{}
}
\value{
无返回值
}
\description{
查询ERP表体input
}
\examples{
erp_arReceivableEntry_select()
}
