% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salOutStock.R
\name{erp_outStockEntry_select}
\alias{erp_outStockEntry_select}
\title{查询ERP,salOutStockEntry}
\usage{
erp_outStockEntry_select(erp_token)
}
\arguments{
\item{erp_token}{}
}
\value{
无返回值
}
\description{
查询ERP,salOutStockEntry
}
\examples{
erp_outStockEntry_select()
}
