% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outStockMerge.R
\name{erp_outStockMerge_input_update}
\alias{erp_outStockMerge_input_update}
\title{拆分表头表体并更新源单内码}
\usage{
erp_outStockMerge_input_update(erp_token)
}
\arguments{
\item{erp_token}{}
}
\value{
无返回值
}
\description{
拆分表头表体并更新源单内码
}
\examples{
erp_outStockMerge_input_update()
}
