% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salOutStock.R
\name{erp_sal_outStockEntry_input_delete}
\alias{erp_sal_outStockEntry_input_delete}
\title{rds_dms_ods_t_sal_outStockEntry_input 清空erp临时表}
\usage{
erp_sal_outStockEntry_input_delete(erp_token)
}
\arguments{
\item{erp_token}{}
}
\value{
无返回值
}
\description{
rds_dms_ods_t_sal_outStockEntry_input 清空erp临时表
}
\examples{
erp_sal_outStockEntry_input_delete()
}
