% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salOutStock.R
\name{erp_salOutStock_input_update}
\alias{erp_salOutStock_input_update}
\title{更新rds_dms_ods_t_sal_outStockEntry_input源单内码}
\usage{
erp_salOutStock_input_update(erp_token)
}
\arguments{
\item{erp_token}{}
}
\value{
无返回值
}
\description{
更新rds_dms_ods_t_sal_outStockEntry_input源单内码
}
\examples{
erp_salOutStock_input_update()
}
