% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salOutStock.R
\name{salOutStock_insert}
\alias{salOutStock_insert}
\title{上传rds_dms_ods_t_sal_outStock表数据}
\usage{
salOutStock_insert(dms_token)
}
\arguments{
\item{dms_token}{}
}
\value{
无返回值
}
\description{
上传rds_dms_ods_t_sal_outStock表数据
}
\examples{
salOutStock_insert()
}
