% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salOutStock.R
\name{salOutStock_select}
\alias{salOutStock_select}
\title{查询salOutStock}
\usage{
salOutStock_select(dms_token, FBillNo)
}
\arguments{
\item{dms_token}{}

\item{FBillNo}{}
}
\value{
无返回值
}
\description{
查询salOutStock
}
\examples{
salOutStock_select()
}
