% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salOutStock.R
\name{salOutStockEntry_insert}
\alias{salOutStockEntry_insert}
\title{上传rds_dms_ods_t_sal_outStockEntry表数据}
\usage{
salOutStockEntry_insert(dms_token)
}
\arguments{
\item{dms_token}{}
}
\value{
无返回值
}
\description{
上传rds_dms_ods_t_sal_outStockEntry表数据
}
\examples{
salOutStockEntry_insert()
}
