% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outStockMerge.R
\name{dms_sal_outStock_view}
\alias{dms_sal_outStock_view}
\title{查询dms 表头更新状态}
\usage{
dms_sal_outStock_view(dms_token, FBillNo)
}
\arguments{
\item{dms_token}{}

\item{FBillNo}{}
}
\value{
无返回值
}
\description{
查询dms 表头更新状态
}
\examples{
dms_sal_outStock_view()
}
