% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salOutStock.R
\name{dms_sal_outStockEntry_input_delete}
\alias{dms_sal_outStockEntry_input_delete}
\title{rds_dms_ods_t_sal_outStockEntry_input清空dms临时表}
\usage{
dms_sal_outStockEntry_input_delete(dms_token)
}
\arguments{
\item{dms_token}{}
}
\value{
无返回值
}
\description{
rds_dms_ods_t_sal_outStockEntry_input清空dms临时表
}
\examples{
dms_sal_outStockEntry_input_delete()
}
